requirements.txt
fastapi
uvicorn
langchain
langchain-text-splitters
chromadb
faiss-cpu
openai
pydantic
python-dotenv
docker
pypdf
unstructured
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
sentence-transformers
ipykernel
pandas
numpy
matplotlib
seaborn
pinecone
python-multipart
jinja2
